requirements.txt
Pillow==10.4.0
schedule==1.2.0
pyautogui==0.9.54
google-api-python-client==2.137.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pyinstaller==6.9.0
cryptography==42.0.5
mss==9.0.1